requirements.txt
pystray
Pillow
customtkinter
pyinstaller
